i2c: chips: corrected a warning for BMA150

+++ drivers/i2c/chips/bma150.c
@@ -191,7 +191,7 @@ static long bma_ioctl(struct file *file, unsigned int cmd,
 	void __user *argp = (void __user *)arg;
 
 	char rwbuf[8] = "";
-	int ret = -1;
+	long ret = -1;
 	short buf[8], temp;
 	int kbuf = 0;
